requirements.txt
# Core Django dependencies
Django==5.1.6
djangorestframework==3.15.2
django-cors-headers==4.6.0
djangorestframework-simplejwt==5.3.0
django-filter==24.3

# Authentication & OAuth
django-allauth==65.2.0
dj-rest-auth==6.0.0

# Database & Models
psycopg2-binary==2.9.10
django-encrypted-model-fields==0.6.5

# AI & Machine Learning
openai==1.58.1
langchain==0.3.9
langchain-openai==0.2.9
scikit-learn==1.5.2
pandas==2.2.3
numpy==1.26.4

# Environment & Configuration
python-dotenv==1.0.1

# Open Banking & Financial APIs
requests==2.32.3
cryptography==44.0.0
pyjwt==2.10.1

# Data Processing & Analytics
pytz==2024.2
python-dateutil==2.9.0

# Development & Testing
pytest==8.3.4
pytest-django==4.9.0

# Optional: Enhanced AI capabilities
# transformers==4.46.3
# torch==2.5.1
# tensorflow==2.18.0

# Optional: Background task processing
# celery==5.4.0
# redis==5.2.1

# Optional: Advanced financial data providers
# plaid-python==17.0.0
# yfinance==0.2.48
